% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocVersion.R
\name{version}
\alias{version}
\title{Bioconductor version}
\usage{
version()
}
\value{
\code{package_version} representing the Bioconductor version in use.
}
\description{
This function reports the Bioconductor version in use, as determined by
the version of the BiocVersion package.
}
\examples{

version()

}
\seealso{
\code{\link{install}} Installs/updates Bioconductor/CRAN packages.
}
\keyword{environment}
